Fix up dir check and rm (SC2115 rule)
See https://github.com/koalaman/shellcheck/wiki/SC2115

+++ server-configs/generate.sh
@@ -18,8 +18,8 @@ mkdir -p "$dest"
 for name in $names; do
     echo "generating configs for $name"
 
-    if [ -d "$name" ]; then
-        rm -r "$dest/$name"
+    if [ -d "$dest/$name" ]; then
+        rm -r "${dest:?}/$name"
     fi
 
     # nginx
